Move reverse from Diff helper to transaction logic
