+++ framework/src/modules/chain/logic/multisignature.js
@@ -17,7 +17,6 @@
 const async = require('async');
 const ByteBuffer = require('bytebuffer');
 const ed = require('../helpers/ed.js');
-const Diff = require('../helpers/diff.js');
 const slots = require('../helpers/slots.js');
 const Bignum = require('../helpers/bignum.js');
 
@@ -35,7 +34,6 @@ __scope.unconfirmedSignatures = {};
  * @see Parent: {@link logic}
  * @requires async
  * @requires bytebuffer
- * @requires helpers/diff
  * @requires helpers/slots
  * @param {Object} scope
  * @param {Object} scope.components
@@ -52,7 +50,7 @@ __scope.unconfirmedSignatures = {};
 class Multisignature {
 	constructor({
 		components: { logger },
-		logic: { transaction, account },
+		logic: { account, transaction },
 		schema,
 		channel,
 	}) {
@@ -62,8 +60,8 @@ class Multisignature {
 		__scope.schema = schema;
 		__scope.channel = channel;
 		__scope.logic = {
-			transaction,
 			account,
+			transaction,
 		};
 
 		// TODO: Add modules to constructor argument and assign accounts to __scope.modules.accounts
@@ -408,7 +406,9 @@ Multisignature.prototype.undoConfirmed = function(
 	cb,
 	tx
 ) {
-	const multiInvert = Diff.reverse(transaction.asset.multisignature.keysgroup);
+	const multiInvert = __scope.logic.transaction.reverse(
+		transaction.asset.multisignature.keysgroup
+	);
 
 	__scope.unconfirmedSignatures[sender.address] = true;
 
@@ -479,7 +479,9 @@ Multisignature.prototype.undoUnconfirmed = function(
 	cb,
 	tx
 ) {
-	const multiInvert = Diff.reverse(transaction.asset.multisignature.keysgroup);
+	const multiInvert = __scope.logic.transaction.reverse(
+		transaction.asset.multisignature.keysgroup
+	);
 
 	__scope.unconfirmedSignatures[sender.address] = false;
