+++ framework/test/mocha/unit/modules/chain/logic/vote.js
@@ -23,7 +23,6 @@ const application = require('../../../../common/application');
 const randomUtil = require('../../../../common/utils/random');
 const modulesLoader = require('../../../../common/modules_loader');
 const ed = require('../../../../../../src/modules/chain/helpers/ed');
-const diff = require('../../../../../../src/modules/chain/helpers/diff');
 const Bignum = require('../../../../../../src/modules/chain/helpers/bignum.js');
 const Transfer = require('../../../../../../src/modules/chain/logic/transfer');
 
@@ -114,7 +113,7 @@ describe('vote', () => {
 	}
 
 	function checkAccountVotes(senderPublicKey, state, votes, action, done) {
-		votes = action === 'apply' ? votes : diff.reverse(votes);
+		votes = action === 'apply' ? votes : transactionLogic.reverse(votes);
 		accountsModule.getAccount(
 			{ publicKey: senderPublicKey },
 			(err, account) => {
@@ -153,6 +152,7 @@ describe('vote', () => {
 					schema: modulesLoader.scope.schema,
 					logic: {
 						account: scope.logic.account,
+						transaction: scope.logic.transaction,
 					},
 				});
